backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.4
django-cors-headers==4.2.0
django-mysql==4.11.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
pika==1.3.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
mysqlclient==2.0.1
djangorestframework-simplejwt==5.2.2
channels==4.0.0
daphne==4.0.0
channels_redis==4.1.0